implemented jwt

+++ app/controllers/tasks_controller.rb
@@ -1,19 +1,28 @@
 class TasksController < ApplicationController
-  before_action :authenticate_staff, only: [:show]
-  before_action :deny_access, only: [:destroy, :create, :show, :update]
+  before_action :authenticate_staff, only: [:index, :show]
+  before_action :deny_staff, only: [:show, :update, :destroy]
+  before_action :deny_access, only: [:show, :create, :update, :destroy]
 
+  # GET /tasks
+  def index
+    if current_admin
+      tasks = Task.all
+    else
+      tasks = @current_staff.tasks
+    end
 
- # GET /tasks
- def index
-  tasks = Task.all
-  render json: tasks, status: :ok
-end
+    render json: tasks, status: :ok
+  end
 
   # GET /tasks/1
   def show
     task = Task.find_by(id: params[:id])
     if task
-      render json: task
+      if current_admin || task.staff_id == @current_staff.id
+        render json: task
+      else
+        render_unauthorized
+      end
     else
       render json: { error: "Task not found" }, status: :not_found
     end
@@ -22,6 +31,8 @@ end
   # POST /tasks
   def create
     task = Task.create(task_params)
+    task.staff_id = @current_staff.id if @current_staff
+
     if task.save
       render json: task, status: :created
     else
@@ -33,10 +44,14 @@ end
   def update
     task = Task.find_by(id: params[:id])
     if task
-      if task.update(task_params)
-        render json: task
+      if current_admin || task.staff_id == @current_staff.id
+        if task.update(task_params)
+          render json: task
+        else
+          render json: { error: task.errors.full_messages }, status: :unprocessable_entity
+        end
       else
-        render json: { error: task.errors.full_messages }, status: :unprocessable_entity
+        render_unauthorized
       end
     else
       render json: { error: "Task not found" }, status: :not_found
@@ -47,8 +62,12 @@ end
   def destroy
     task = Task.find_by(id: params[:id])
     if task
-      task.destroy
-      head :no_content
+      if current_admin || task.staff_id == @current_staff.id
+        task.destroy
+        head :no_content
+      else
+        render_unauthorized
+      end
     else
       render json: { error: "Task not found" }, status: :not_found
     end
@@ -60,12 +79,20 @@ end
     params.permit(:task_name, :assigned_to, :managed_by, :project_id, :staff_id)
   end
 
+  def deny_staff
+    render_unauthorized unless current_admin
+  end
+
   def deny_access
-    render_unauthorized unless authenticate_admin
+    render_unauthorized unless current_admin || (current_staff && task_belongs_to_current_staff?)
+  end
+
+  def task_belongs_to_current_staff?
+    task = Task.find_by(id: params[:id])
+    task && task.staff_id == @current_staff.id
   end
 
   def render_unauthorized
     render json: { error: 'Unauthorized' }, status: :unauthorized
   end
-
 end
